refactor(emit): extract value emission helpers to dedicated module
Extract value emission and boxing/unboxing functions from emit.rs (~303 lines)
to value_emission.rs module (333 lines):
- emit_operands: Emit operands with local.get and nil handling
- emit_operands_with_boxing: Emit operands with type boxing for generic calls
- emit_value: Emit single value with fallback for block args
- emit_boxing: Box concrete types (Int/Float) to anyref
- emit_unboxing: Unbox anyref to concrete types
- infer_call_result_type: Type inference for generic function calls

This reduces emit.rs from 3,938 to 3,635 lines (~303 lines moved).
The new value_emission module is 333 lines.
